+++ src/components/ProductTile/ProductTileValue.test.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+import { shallow } from 'enzyme';
+import ProductTileValue from './ProductTileValue';
+
+const props = {
+    value: 'werwer',
+    isServer: false,
+};
+
+describe('<ProductTileValue />', () => {
+    it('it renders ProductTileValue', () => {
+        const wrapper = shallow(<ProductTileValue {...props} hAlign="center" />);
+        expect(wrapper).toMatchSnapshot();
+    });
+});